Test every region; cleanup
