+++ labor_abm_canada/regions/regions.py
@@ -1,0 +1,17 @@
+from enum import StrEnum
+
+
+class Regions(StrEnum):
+    Alberta_A = "Alberta.a"
+    British_Columbia_A = "British Columbia.a"
+    Manitoba_A = "Manitoba.a"
+    New_Brunswick_A = "New Brunswick.a"
+    Newfoundland_and_Labrador_A = "Newfoundland and Labrador.a"
+    Nova_Scotia_A = "Nova Scotia.a"
+    Ontario_A = "Ontario.a"
+    Ontario_B = "Ontario.b"
+    Prince_Edward_Island_A = "Prince Edward Island.a"
+    Quebec_A = "Quebec.a"
+    Quebec_B = "Quebec.b"
+    Saskatchewan_A = "Saskatchewan.a"
+    National = "National"
