+++ requirements.txt
@@ -4,66 +4,69 @@
 #
 #    pip-compile
 #
-annotated-types==0.7.0
+annotated-types~=0.7.0
     # via pydantic
-et-xmlfile==1.1.0
+et-xmlfile~=1.1.0
     # via openpyxl
-filelock==3.15.4
+filelock~=3.15.4
     # via torch
-fsspec==2024.6.1
+fsspec~=2024.6.1
     # via torch
-iniconfig==2.0.0
+iniconfig~=2.0.0
     # via pytest
-jinja2==3.1.4
+jinja2~=3.1.4
     # via torch
-markupsafe==2.1.5
+markupsafe~=2.1.5
     # via jinja2
-mpmath==1.3.0
+mpmath~=1.3.0
     # via sympy
-networkx==3.3
+networkx~=3.3
     # via
     #   -r requirements.in
     #   torch
-numpy==2.0.0
+numpy~=2.0.0
     # via
     #   -r requirements.in
     #   pandas
     #   scipy
-openpyxl==3.1.5
+openpyxl~=3.1.5
     # via -r requirements.in
-packaging==24.1
+packaging~=24.1
     # via pytest
-pandas==2.2.2
+pandas~=2.2.2
     # via -r requirements.in
-pluggy==1.5.0
+pluggy~=1.5.0
     # via pytest
-pydantic==2.8.2
+pydantic~=2.8.2
     # via -r requirements.in
-pydantic-core==2.20.1
+pydantic-core~=2.20.1
     # via pydantic
-pytest==8.2.2
+pytest~=8.2.2
     # via -r requirements.in
-python-dateutil==2.9.0.post0
+python-dateutil~=2.9.0.post0
     # via pandas
-pytz==2024.1
+pytz~=2024.1
     # via pandas
-pyyaml==6.0.1
+pyyaml~=6.0.1
     # via -r requirements.in
-scipy==1.14.0
+scipy~=1.14.0
     # via -r requirements.in
-six==1.16.0
+six~=1.16.0
     # via python-dateutil
-sympy==1.12.1
+sympy~=1.12.1
     # via torch
-torch==2.3.1
+torch~=2.3.1
     # via -r requirements.in
-typing-extensions==4.12.2
+typing-extensions~=4.12.2
     # via
     #   pydantic
     #   pydantic-core
     #   torch
-tzdata==2024.1
+tzdata~=2024.1
     # via pandas
 
 # The following packages are considered to be unsafe in a requirements file:
 # setuptools
+
+setuptools~=70.0.0
+tqdm~=4.66.4
