+++ test_labor_abm/test_data/model_params.yaml
@@ -1,0 +1,10 @@
+# a0: -0.0105444443610075
+# a1: 0.216086177536022
+# a2: 0.1997362327000295
+# a3: 0.2120885748885414
+# a4: 0.1609228338413122
+# a5: 0.2089889214977157
+separation_rate: 0.0203842505225833
+opening_rate: 0.0185538914648699
+adaptation_rate_u: 0.3960846582993839
+adaptation_rate_v: 0.1744454655908199
