+++ test_labor_abm/test_model/test_run.py
@@ -1,10 +1,13 @@
+import pytest
 import torch
 
-from labor_abm import LaborABM
+from labor_abm_canada.model.labor_abm import LabourABM
+from labor_abm_canada.regions.regions import Regions
+from labor_abm_canada.runner import run_model
 
 
 class TestModel:
-    def test_run(self, twonode_model: LaborABM):
+    def test_run(self, twonode_model: LabourABM):
         twonode_model.run_model()
 
         aggregates = twonode_model.aggregates
@@ -13,3 +16,11 @@ class TestModel:
         assert torch.all(aggregates.total_vacancies >= 0)
         assert torch.all(aggregates.total_employment >= 0)
         assert torch.all(aggregates.total_demand >= 0)
+
+
+class TestRunner:
+
+    @pytest.mark.parametrize("region", Regions)
+    def test_run(self, default_scenario_path, default_runner_config, region: Regions):
+        run_model(scenario_filename=default_scenario_path, region=region, burn_in=0)
+        assert True
